feat: move across admin trackgroup endoints
