+++ src/controllers/index.mjs
@@ -68,6 +68,12 @@ import usersId from './users/routes/{id}/index.js'
 import usersIdPlaylists from './users/routes/{id}/playlists.js'
 import adminUsers from './user/admin/users/index.js'
 import adminUsersId from './user/admin/users/{id}.js'
+import adminTrackgroups from './user/admin/trackgroups/index.js'
+import adminTrackgroupId from './user/admin/trackgroups/{id}/index.js'
+import adminTrackgroupSettings from './user/admin/trackgroups/{id}/settings.js'
+import adminTrackgroupItems from './user/admin/trackgroups/{id}/items/index.js'
+import adminTrackgroupItemsAdd from './user/admin/trackgroups/{id}/items/add.js'
+import adminTrackgroupItemsRemove from './user/admin/trackgroups/{id}/items/remove.js'
 
 import { apiRoot as root } from '../constants.js'
 
@@ -85,7 +91,6 @@ initialize({
     { path: `${root}/tracks/latest`, module: tracksRoutesLatest },
     { path: `${root}/tracks/{id}`, module: tracksId },
 
-    // FIXME: This one is broken cause it relies on elastic search.
     { path: `${root}/tag/{tag}`, module: tagRoutesTag },
 
     { path: `${root}/artists/featured`, module: artistsFeatured },
@@ -101,7 +106,6 @@ initialize({
     { path: `${root}/labels/{id}/releases`, module: labelReleases },
     { path: `${root}/labels/{id}/artists`, module: labelArtists },
 
-    // FIXME: This one is broken cause it relies on elastic search.
     { path: `${root}/search`, module: search },
 
     // FIXME: Not entirely clear on what the point of this route is
@@ -161,6 +165,13 @@ initialize({
 
     { path: `${root}/user/admin/users`, module: adminUsers },
     { path: `${root}/user/admin/users/{id}`, module: adminUsersId },
+
+    { path: `${root}/user/admin/trackgroups`, module: adminTrackgroups },
+    { path: `${root}/user/admin/trackgroups/{id}`, module: adminTrackgroupId },
+    { path: `${root}/user/admin/trackgroups/{id}/settings`, module: adminTrackgroupSettings },
+    { path: `${root}/user/admin/trackgroups/{id}/items`, module: adminTrackgroupItems },
+    { path: `${root}/user/admin/trackgroups/{id}/items/add`, module: adminTrackgroupItemsAdd },
+    { path: `${root}/user/admin/trackgroups/{id}/items/remove`, module: adminTrackgroupItemsRemove },
 
     {
       path: `${root}/apiDocs`, module: apiDocs
